+++ app/helpers/get-with-default.js
@@ -1,0 +1,1 @@
+export {default, getWithDefault} from 'ember-frost-list/helpers/get-with-default'