+++ app/utils/get-component-name.js
@@ -1,0 +1,1 @@
+export {default} from 'ember-frost-list/utils/get-component-name'