+++ tests/dummy/app/router.js
@@ -15,6 +15,7 @@ Router.map(function () {
   this.route('infinite')
   this.route('paged')
   this.route('size')
+  this.route('typed')
 })
 
 export default Router